requirements.txt
asgiref==3.6.0
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.2.3
django-admin-decorators==0.1
django-admin-interface==0.26.0
django-admin-list-filter-dropdown==1.0.3
django-admin-multi-select-filter==1.2.1
django-admin-searchable-dropdown==1.2
django-colorfield==0.9.0
django-dynamic-raw-id==3.0
django-filebrowser==4.0.2
django-grappelli==3.0.4
django-import-export==3.0.2
django-jalali==6.0.1
django-jalali-date==1.0.1
django-more-admin-filters==1.7
django-render-partial==0.4
django-salmonella==1.2.1
django-social-share==2.3.0
django-summernote==0.8.20.0
django-tinymce==3.5.0
django-user-language-middleware==0.0.3
docutils==0.19
et-xmlfile==1.1.0
filelock==3.9.0
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
jdatetime==4.1.0
jeepney==0.8.0
jsmin==3.0.1
keyring==23.13.1
markdown-it-py==2.1.0
MarkupPy==1.14
mdurl==0.1.2
more-itertools==9.0.0
odfpy==1.4.1
openpyxl==3.1.0
packaging==23.0
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.14.0
pyproject_api==1.5.0
python-slugify==8.0.1
pytz==2022.7.1
pywin32-ctypes==0.2.2
PyYAML==6.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.1
SecretStorage==3.3.3
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
text-unidecode==1.3
tomli==2.0.1
tox==4.4.4
twine==4.0.2
tzdata==2023.3
urllib3==1.26.14
virtualenv==20.18.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.12.0
